refactor tree.py and tree_async.py
