+++ mcts/tree.py
@@ -9,22 +9,20 @@ import numpy as np
 import torch
 
 from board.constant import PASS, RESIGN
-from board.coordinate import Coordinate
 from board.go_board import GoBoard, copy_board
 from board.stone import Stone
 from common.print_console import print_err
-from nn.feature import generate_input_planes
-from nn.network.dual_net import DualNet
-from nn.tentative_policy import get_tentative_policy
 from mcts.batch_data import BatchQueue
 from mcts.constant import NOT_EXPANDED, PLAYOUTS, NN_BATCH_SIZE, \
     MAX_CONSIDERED_NODES, RESIGN_THRESHOLD, MCTS_TREE_SIZE
 from mcts.sequential_halving import get_candidates_and_visit_pairs
-from mcts.node import MCTSNode
+from mcts.tree_base import MCTSTreeBase
 from mcts.time_manager import TimeControl, TimeManager
-from mcts.dump import dump_mcts_to_json
+from nn.feature import generate_input_planes
+from nn.network.dual_net import DualNet
 
-class MCTSTree: # pylint: disable=R0902
+
+class MCTSTree(MCTSTreeBase): # pylint: disable=R0902
     """モンテカルロ木探索の実装クラス。
     """
     def __init__(self, network: DualNet, tree_size: int=MCTS_TREE_SIZE, \
@@ -36,15 +34,9 @@ class MCTSTree: # pylint: disable=R0902
             tree_size (int, optional): 木を構成するノードの最大個数。デフォルトは65536。
             batch_size (int, optional): ニューラルネットワークの前向き伝搬処理のミニバッチサイズ。デフォルトはNN_BATCH_SIZE。
         """
-        self.node = [MCTSNode() for i in range(tree_size)]
-        self.num_nodes = 0
-        self.root = 0
+        super().__init__(tree_size=tree_size, batch_size=batch_size, cgos_mode=cgos_mode)
         self.network = network
         self.batch_queue = BatchQueue()
-        self.current_root = 0
-        self.batch_size = batch_size
-        self.cgos_mode = cgos_mode
-        self.to_move = Stone.BLACK
 
 
     def _initialize_search(self, board: GoBoard, color: Stone) -> None:
@@ -175,9 +167,10 @@ class MCTSTree: # pylint: disable=R0902
             sys.stdout.flush()
 
 
-    def search_with_callback(self, board: GoBoard, color: Stone, callback: Callable[[List[Tuple[int, int]]], bool]) -> None:
+    def search_with_callback(self, board: GoBoard, color: Stone, \
+        callback: Callable[[List[Tuple[int, int]]], bool]) -> None:
         """探索を実行し、探索系列をコールバック関数へ渡す動作をくり返す。
-コールバック関数の戻り値が真になれば終了する。
+        コールバック関数の戻り値が真になれば終了する。
         Args:
             board (GoBoard): 現在の局面情報。
             color (Stone): 現局面の手番の色。
@@ -243,32 +236,6 @@ class MCTSTree: # pylint: disable=R0902
         else:
             next_node_index = self.node[current_index].get_child_index(next_index)
             self.search_mcts(board, color, next_node_index, path)
-
-
-    def expand_node(self, board: GoBoard, color: Stone) -> int:
-        """ノードを展開する。
-
-        Args:
-            board (GoBoard): 現在の局面情報。
-            color (Stone): 現在の手番の色。
-        """
-        node_index = self.num_nodes
-        tree_size = len(self.node)
-        if node_index >= tree_size:
-            self.node.extend([MCTSNode() for i in range(tree_size)])
-            sys.stderr.write(f"Tree is full. Allocate new space {tree_size} -> {len(self.node)}\n")
-
-        candidates = board.get_all_legal_pos(color)
-        candidates = [candidate for candidate in candidates \
-            if (board.check_self_atari_stone(candidate, color) < 7) \
-                and not board.is_complete_eye(candidate, color)]
-        candidates.append(PASS)
-
-        policy = get_tentative_policy(candidates)
-        self.node[node_index].expand(policy)
-
-        self.num_nodes += 1
-        return node_index
 
 
     def process_mini_batch(self, board: GoBoard, use_logit: bool=False): # pylint: disable=R0914
@@ -421,87 +388,3 @@ class MCTSTree: # pylint: disable=R0902
                 self.node[current_index].set_child_index(next_index, child_index)
             next_node_index = self.node[current_index].get_child_index(next_index)
             self.search_sequential_halving(board, color, next_node_index, path, count_threshold)
-
-    def get_root(self) -> MCTSNode:
-        """木のルートを返す。
-
-        Returns:
-            MCTSNode: モンテカルロ木探索で使用する木のルート。
-        """
-        return self.node[self.current_root]
-
-    def get_pv_lists(self, root: MCTSNode, coord: Coordinate) -> Dict[str, List[str]]:
-        """探索した手の最善応手系列を取得する。
-
-        Args:
-            coordinate (Coordinate): 座標変換処理インスタンス。
-
-        Returns:
-            Dict[str, List[str]]: 各手の最善応手系列を記録した辞書。
-        """
-        pv_dict: Dict[str, List[str]] = {}
-
-        for i in range(root.num_children):
-            if root.children_visits[i] > 0:
-                pv_list = self.get_best_move_sequence([root.action[i]], root.children_index[i])
-                pv_dict[coord.convert_to_gtp_format(root.action[i])] = \
-                    [coord.convert_to_gtp_format(pv) for pv in pv_list]
-
-        return pv_dict
-
-    def get_best_move_sequence(self, pv_list: List[int], index: int) -> List[int]:
-        """最善応手系列を取得する。
-
-        Args:
-            pv_list (List[str]): 今までの経路の最善応手系列。
-            index (int): ノードのインデックス。
-
-        Returns:
-            List[str]: 最善応手系列。
-        """
-        node = self.node[index]
-
-        if node.node_visits == 0:
-            return pv_list
-
-        next_index = node.get_child_index(node.get_best_move_index())
-        next_action = node.get_best_move()
-        pv_list.append(next_action)
-
-        if next_index == NOT_EXPANDED:
-            return pv_list
-
-        return self.get_best_move_sequence(pv_list, next_index)
-
-
-    def dump_to_json(self, board: GoBoard, superko: bool) -> str:
-        """MCTSの状態を表すJSON文字列を返す。
-
-        Args:
-            board (GoBoard): 現在の碁盤。
-            superko (bool): 超劫判定の有効化。
-
-        Returns:
-            str: MCTSの状態を表すJSON文字列。
-        """
-        return dump_mcts_to_json(self.to_dict(), board, superko)
-
-
-    def to_dict(self) -> Dict[str, Any]:
-        """ツリーの状態を辞書化して返す。
-
-        Returns:
-            Dict[str, Any]: ツリーの状態を表す辞書。
-        """
-        state = {
-            "node": [self.node[i].to_dict() for i in range(self.num_nodes)],
-            "num_nodes": self.num_nodes,
-            "root": self.root,
-            #"network": self.network,  # ダンプに含めない
-            #"batch_queue": self.batch_queue,  # ダンプに含めない
-            "current_root": self.current_root,
-            "batch_size": self.batch_size,
-            "cgos_mode": self.cgos_mode,
-            "to_move": 'black' if self.to_move == Stone.BLACK else 'white',
-        }
-        return state
